Move DashboardData to the correct location
