+++ features/dashboard/src/main/kotlin/com/boswelja/truemanager/dashboard/ui/overview/cards/MemoryCard.kt
@@ -27,7 +27,7 @@ import com.boswelja.capacity.Capacity
 import com.boswelja.capacity.Capacity.Companion.gigabytes
 import com.boswelja.capacity.CapacityUnit
 import com.boswelja.truemanager.dashboard.R
-import com.boswelja.truemanager.dashboard.ui.DashboardData
+import com.boswelja.truemanager.dashboard.ui.overview.DashboardData
 import com.boswelja.truemanager.dashboard.ui.overview.cards.common.CardListItem
 import com.boswelja.truemanager.dashboard.ui.overview.cards.common.DashboardCard
