+++ features/dashboard/src/main/kotlin/com/boswelja/truemanager/dashboard/ui/overview/cards/SystemInformationCard.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.boswelja.truemanager.dashboard.R
-import com.boswelja.truemanager.dashboard.ui.DashboardData
+import com.boswelja.truemanager.dashboard.ui.overview.DashboardData
 import com.boswelja.truemanager.dashboard.ui.overview.cards.common.CardListItem
 import com.boswelja.truemanager.dashboard.ui.overview.cards.common.DashboardCard
 import kotlinx.coroutines.delay
